Controlador, validadores, mapper OnSaleDTO

+++ Microservicio-Pedidos/src/main/kotlin/biques/dam/es/models/OrderLine.kt
@@ -2,11 +2,10 @@ package biques.dam.es.models
 
 import kotlinx.serialization.Contextual
 import kotlinx.serialization.Serializable
-import kotlinx.serialization.Serializer
 import org.bson.codecs.pojo.annotations.BsonId
 import org.litote.kmongo.Id
 import org.litote.kmongo.newId
-import serializers.UUIDSerializer
+import biques.dam.es.serializers.UUIDSerializer
 import java.util.UUID
 
 @Serializable
@@ -18,8 +17,8 @@ data class OrderLine(
     //@Contextual
     //val product: Id<OnSale>,
     val amount: Int,
-    val price: Float,
-    val total: Int,
+    val price: Double,
+    val total: Double,
     //@Contextual
     //val employee: Id<User>
     )
